swdl: ignore SIGPIPE, enable printing errors on tar failure
If we give a bad boot_tar* part type, tar errors out and closes the
pipe, subsequent writes to our end of the pipe raise a SIGPIPE which by
default exits immediately.  Since we always check the return code of
write, we'll detect that error (EPIPE) and handle it sanely.