mknImage: create: suppress stringop-truncation warning
GCC warns that the strncpy for hdr.name may truncate, which is true, but
OK for mknImage since a null terminator isn't actually required (and
we're careful to printf it with an explicit size %.*s).

Use pragmas to locally suppress that error rather than doing so
globally. Compile with -Wno-unknown-warning-option so that clang doesn't
complain about the unknown -Wstring-op-truncation.