init: add systemd service to save last boot timestamp
It's trickier than expected to get a service to *start* during the
shutdown process, but (as suggested on forums) it does work to make
a oneshot service with a useful ExecStop.

It's also hard because we need /boot to be mounted read-write still, and
most of the shutdown targets or /usr/lib/systemd/system-shutdown scripts
run very late after everything's already been unmounted.